refactor: class labeling 수정
