+++ src/main/java/christmas/domain/discount/strategies/Gift.java
@@ -1,13 +1,13 @@
 package christmas.domain.discount.strategies;
 
-import static christmas.domain.discount.constants.DiscountInfo.DEFAULT_GIFT_QUANTITY;
-import static christmas.domain.discount.constants.DiscountInfo.GIFT_DISCOUNT;
+import static christmas.domain.discount.constants.DiscountLabel.GIFT_DISCOUNT;
 
 import christmas.domain.Orders;
 import christmas.domain.discount.Discount;
 import christmas.domain.vo.Product;
 
 public class Gift implements Discount {
+    private static final String DEFAULT_GIFT_QUANTITY = "1개";
     private final boolean isGiftAvailable;
 
     public Gift(boolean isGiftAvailable) {
