+++ src/test/java/christmas/domain/discount/strategies/SpecialDiscountTest.java
@@ -2,7 +2,7 @@ package christmas.domain.discount.strategies;
 
 import static christmas.domain.discount.constants.DiscountNumeric.SPECIAL;
 import static christmas.domain.discount.constants.DiscountNumeric.SPECIAL_DISCOUNT_PRICE;
-import static christmas.domain.discount.constants.DiscountInfo.SPECIAL_DISCOUNT;
+import static christmas.domain.discount.constants.DiscountLabel.SPECIAL_DISCOUNT;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import christmas.domain.OrderProduct;
